Parce que les commits via commande marche pu

+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: rjakubec <rjakubec@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/14 12:44:37 by rsk               #+#    #+#             */
-/*   Updated: 2017/11/16 15:49:32 by rjakubec         ###   ########.fr       */
+/*   Updated: 2017/11/16 19:55:17 by rjakubec         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strlen(const char *s)
+size_t		ft_strlen(const char *s)
 {
 	int i;
